Fix default binding behaviour
On some platforms, binding to to both "::" and "0.0.0.0" at the same
time is not allowed because "::" may already accepts ipv4 connections.
For other platforms, binding to both is necessary to support ipv4 and
ipv6.

This platform-specific behaviour is due to the variation in the
default value for the socket option "IPV6_ONLY".

Fix this by always setting the "IPv6_ONLY" sockopt to true!